tests/wpt/web-platform-tests/fullscreen/api/document-fullscreen-enabled.html

<!DOCTYPE html>
<title>Document.fullscreenEnabled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="blank.html"></iframe>
<iframe allowfullscreen src="blank.html"></iframe>
<script>
    async_test(function(t)
{
    assert_true(document.fullscreenEnabled, "top-level document");

    var iframes = document.getElementsByTagName("iframe");

    iframes[0].onload = t.step_func_done(function() {
        assert_false(iframes[0].contentDocument.fullscreenEnabled, "iframe without allowfullscreen");
        iframes[1].onload = t.step_func_done(function() {
            assert_true(iframes[1].contentDocument.fullscreenEnabled, "iframe with allowfullscreen");
        });
    });
});
</script>
